+++ clodoo/__main__.py
@@ -16,7 +16,7 @@ except ImportError:
     from zerobug import Z0BUG
 
 
-__version__ = "0.3.31.2"
+__version__ = "0.3.31.3"
 STS_FAILED = 1
 STS_SUCCESS = 0