+++ clodoo/export_db_model.py
@@ -22,7 +22,7 @@ except ImportError:
 # import pdb
 
 
-__version__ = "0.3.31.2"
+__version__ = "0.3.31.3"
 
 
 CACHE = {}