+++ clodoo/odoo_skin.sh
@@ -57,7 +57,7 @@ fi
 TESTDIR=$(findpkg "" "$TDIR . .." "tests")
 RUNDIR=$(readlink -e $TESTDIR/..)
 
-__version__=0.3.31.2
+__version__=0.3.31.3
 
 
 get_odoo_service_name() {