+++ clodoo/tests/__pytest_transodoo.py
@@ -12,7 +12,7 @@
 # import pytest
 from clodoo import transodoo
 
-__version__ = '0.3.31.2'
+__version__ = '0.3.31.3'
 
 
 class TestClass: