+++ clodoo/tests/clodoo_test_02.py
@@ -29,7 +29,7 @@ except ImportError:
     import clodoo
 
 
-__version__ = "0.3.31.2"
+__version__ = "0.3.31.3"
 
 
 MODULE_ID = 'clodoo'