+++ clodoo/update_coa.py
@@ -15,7 +15,7 @@ except ImportError:
 # import pdb
 
 
-__version__ = "0.3.31.2"
+__version__ = "0.3.31.3"
 
 
 msg_time = time.time()