+++ z0bug_odoo/__main__.py
@@ -20,7 +20,7 @@ except ImportError:
 standard_library.install_aliases()
 
 
-__version__ = "1.0.2.1"
+__version__ = "1.0.2.2"
 STS_FAILED = 1
 STS_SUCCESS = 0