+++ z0bug_odoo/tests/test_z0bug_odoo.py
@@ -13,7 +13,7 @@ import sys
 from zerobug import Z0test
 
 
-__version__ = "1.0.2.1"
+__version__ = "1.0.2.2"
 
 MODULE_ID = 'z0bug_odoo'