+++ zerobug/__main__.py
@@ -19,7 +19,7 @@ except ImportError:
 standard_library.install_aliases()
 
 
-__version__ = '1.0.1.1'
+__version__ = '1.0.1.2'
 STS_FAILED = 1
 STS_SUCCESS = 0