+++ zerobug/tests/zerobug_test_05.py
@@ -11,7 +11,7 @@ import sys
 from zerobug import Z0BUG
 
 
-__version__ = "1.0.1.1"
+__version__ = "1.0.1.2"
 
 MODULE_ID = 'zerobug'
 TEST_FAILED = 1