+++ zerobug/tests/zerobug_test_08.py
@@ -13,7 +13,7 @@ import sys
 from zerobug import Z0BUG
 from zerobug import Z0testOdoo
 
-__version__ = "1.0.1.1"
+__version__ = "1.0.1.2"
 
 MODULE_ID = 'zerobug'
 TEST_FAILED = 1